requirements.txt
inflect
python-dateutil==2.5.3
python-telegram-bot==7.0.1
requests
unittest2==1.1.0
emoji==0.3.9
peewee==2.8.5
flask
flask-autodoc
flask-security
flask-admin==1.5.0
wtf-peewee==0.2.6
